Minor grammar and spelling fixes
All edits are within the 'About this project' section.
I corrected the spelling of 'conceptes' to 'concepts'. 
I changed 'catalogue of handbook'  to 'catalogue of handbooks'.
I changed 'could locate practical syntax' to 'can locate practical syntax'. 
That last change is necessary because it should match with the 'can' earlier in the sentence. 

Super cool project! Keep up the good work.